Remove module load during freeze
It is not clear why we need to load the module during freeze.  The module local state is stored in globals, so I don't see why we need this.

I think our original understanding of where module local state was stored was incorrect, which probably lead to this code.

If we remove this code, then we can also do not seem to need the immutable keys on the module?