Modify freeze to reuse tp_traverse
This changes the structure of freezing a function to be a two step operation,
the first step is to clone the relevant part of the globals and builtins that a function uses.

The second part is to then traverse using the standard traverse functions.

This overall simplifies the code, and makes it easier to integrate the SCC code.